pyproject.toml
[project]
name = "otterai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "python-dotenv>=1.2.1",
    "requests",
    "requests_toolbelt",
]

[project.scripts]
otter = "otterai.cli:main"

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "black"]
